MaxKernel/requirements.txt
torch
google-cloud-bigquery
langchain-google-community
tensorflow
transformers
jax[tpu]
matplotlib
pandas
protobuf
anthropic[vertex]